More charset.h cleanups
Now charset.h is just character set conversions
